--- a/gcs/go.mod
+++ b/gcs/go.mod
@@ -7,6 +7,6 @@ require (
 	github.com/dchest/siphash v1.2.1
 	github.com/decred/dcrd/blockchain/stake v1.1.0
 	github.com/decred/dcrd/chaincfg/chainhash v1.0.1
-	github.com/decred/dcrd/txscript v1.0.2
+	github.com/decred/dcrd/txscript/v2 v2.0.0
 	github.com/decred/dcrd/wire v1.2.0
 )