--- a/gcs/go.mod
+++ b/gcs/go.mod
@@ -5,7 +5,7 @@ go 1.11
 require (
 	github.com/dchest/blake256 v1.0.0
 	github.com/dchest/siphash v1.2.1
-	github.com/decred/dcrd/blockchain/stake v1.1.0
+	github.com/decred/dcrd/blockchain/stake/v2 v2.0.0
 	github.com/decred/dcrd/chaincfg/chainhash v1.0.1
 	github.com/decred/dcrd/txscript/v2 v2.0.0
 	github.com/decred/dcrd/wire v1.2.0