This updates the `gcs` dependencies and serves as a base for `gcs/v1.1.0`.

The updated direct dependencies in this PR are as follows:

- github.com/decred/dcrd/blockchain/stake/v2@v2.0.0
- github.com/decred/dcrd/txscript/v2@v2.0.0

The full list of updated direct dependencies since the previous `gcs/v1.0.2` release are as follows:

- github.com/decred/dcrd/blockchain/stake/v2@v2.0.0
- github.com/dchest/siphash@v1.2.1
- github.com/decred/dcrd/txscript/v2@v2.0.0
- github.com/decred/dcrd/wire@v1.2.0

It also optimizes the `AddSigScript` method on the `blockcf.Entries` type by making use of the zero-alloc tokenizer now available in the `txscript/v2` module.

The following is a before and after comparison of analyzing typical standard signature script to redeem a pay-to-pubkey-hash:

```
benchmark               old ns/op    new ns/op    delta
---------------------------------------------------------
BenchmarkAddSigScript   494          374          -24.29%

benchmark               old allocs   new allocs   delta
---------------------------------------------------------
BenchmarkAddSigScript   3            2            -33.33%

benchmark               old bytes    new bytes    delta
---------------------------------------------------------
BenchmarkAddSigScript   128          80           -37.50%
```